Revert back to addon.py
